Update gitignore to ignore vscode, add Word Break
Sometimes when running Python in Visual Studio Code it will forget
the which Python interpreter to use and needs to update the local
directory with a .vscode folder that has a launch.json file with
runtime configuration information. We're going to ignore those files.
Also adding the solution for Word Break in here.